requirements.txt
annotated-types==0.7.0
anyio==4.12.1
babel==2.18.0
beautifulsoup4==4.14.3
certifi==2026.2.25
chardet==5.2.0
charset-normalizer==3.4.6
courlan==1.3.2
cssselect==1.3.0
dateparser==1.2.2
exceptiongroup==1.3.1
faiss-cpu==1.13.0
filelock==3.19.1
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.4.2
htmldate==1.9.4
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
joblib==1.5.3
jusText==3.0.2
lxml==6.0.2
lxml_html_clean==0.4.4
Markdown==3.9
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
ollama==0.6.1
packaging==26.0
pillow==11.3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
readability-lxml==0.8.4.1
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
six==1.17.0
soupsieve==2.8.3
sympy==1.14.0
threadpoolctl==3.6.0
tld==0.13.2
tokenizers==0.22.2
torch==2.8.0
tqdm==4.67.3
trafilatura==2.0.0
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
zipp==3.23.0
